# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Login input or transport is invalid."
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Refresh credential transport is invalid."
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Current device session revoked and Web refresh cookie cleared."
          },
          "400": {
            "description": "Logout does not accept a user or session target."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationAcceptedDto"
                }
              }
            }
          },
          "400": {
            "description": "Registration input or transport is invalid."
          },
          "409": {
            "description": "The username is already taken."
          }
        },
        "summary": "",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/households/{householdId}/events": {
      "post": {
        "operationId": "createEvent",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "events"
        ]
      },
      "get": {
        "operationId": "listEvents",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "ISO date string (inclusive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "ISO date string (inclusive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recurring",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expandRecurring",
            "required": false,
            "in": "query",
            "description": "Generate recurring events for a bounded calendar range (at most 366 days).",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/households/{householdId}/events/{eventId}": {
      "get": {
        "operationId": "getEvent",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "events"
        ]
      },
      "put": {
        "operationId": "updateEvent",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "events"
        ]
      },
      "delete": {
        "operationId": "deleteEvent",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "events"
        ]
      }
    },
...
  },
  "info": {
    "title": "Muchakucha Zwei API",
    "description": "Versioned household collaboration API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "family_member",
            "minLength": 3,
            "maxLength": 64,
            "description": "Allows Unicode lowercase expansion of a registered username."
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "writeOnly": true
          },
          "platform": {
            "type": "string",
            "enum": [
              "native",
              "web"
            ]
          }
        },
        "required": [
          "username",
          "password",
          "platform"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string",
            "description": "Native-only refresh credential. Web responses omit this property."
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Native-only opaque refresh credential.",
            "minLength": 1
          }
        }
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string",
            "description": "Native-only rotated refresh credential. Web responses omit this property."
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "family_member",
            "minLength": 3,
            "maxLength": 32,
            "description": "Use username and confirmPassword for immediate registration."
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "writeOnly": true
          },
          "confirmPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "writeOnly": true,
            "description": "Required and must match password for username registration."
          },
          "platform": {
            "type": "string",
            "enum": [
              "native",
              "web"
            ]
          }
        },
        "required": [
          "username",
          "password",
          "confirmPassword",
          "platform"
        ]
      },
      "RegistrationAcceptedDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "REGISTRATION_ACCEPTED"
            ]
          },
          "accessToken": {
            "type": "string",
            "description": "Access credential issued immediately for username registration."
          },
          "refreshToken": {
            "type": "string",
            "description": "Native-only refresh credential for username registration."
          }
        },
        "required": [
          "code",
          "accessToken"
        ]
      },
...
      "RecurrenceDto": {
        "type": "object",
        "properties": {
          "freq": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "yearly"
            ]
          },
          "interval": {
            "type": "number",
            "minimum": 1,
            "maximum": 52,
            "default": 1
          },
          "byWeekday": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "startsOn": {
            "type": "string",
            "example": "2026-08-13"
          },
          "endsOn": {
            "type": "string",
            "example": "2026-12-31"
          },
          "count": {
            "type": "number",
            "minimum": 1,
            "maximum": 1000
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Shanghai"
          },
          "startTimeLocal": {
            "type": "string",
            "example": "08:30"
          },
          "durationMinutes": {
            "type": "number",
            "minimum": 0,
            "maximum": 1440
          }
        },
        "required": [
          "freq",
          "startsOn",
          "timezone"
        ]
      },
...
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Event title",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "description": "Optional description"
          },
          "startTime": {
            "type": "string",
            "description": "Start time in ISO 8601 format"
          },
          "endTime": {
            "type": "string",
            "description": "End time in ISO 8601 format"
          },
          "allDay": {
            "type": "boolean",
            "description": "Whether this is an all-day event",
            "default": false
          },
          "location": {
            "type": "string",
            "description": "Optional location",
            "maxLength": 255
          },
          "recurrence": {
            "description": "重复规则；省略即普通一次性事件",
            "allOf": [
              {
                "$ref": "#/components/schemas/RecurrenceDto"
              }
            ]
          }
        },
        "required": [
          "title",
          "startTime",
          "endTime"
        ]
      },
      "RecurrenceResponseDto": {
        "type": "object",
        "properties": {
          "updatedAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "freq": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "yearly"
            ]
          },
          "interval": {
            "type": "number"
          },
          "byWeekday": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "startsOn": {
            "type": "string"
          },
          "endsOn": {
            "type": "object",
            "nullable": true
          },
          "count": {
            "type": "object",
            "nullable": true
          },
          "timezone": {
            "type": "string"
          },
          "materializedThrough": {
            "type": "object",
            "nullable": true
          },
          "startTimeLocal": {
            "type": "object",
            "nullable": true
          },
          "durationMinutes": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "updatedAt",
          "id",
          "freq",
          "interval",
          "byWeekday",
          "startsOn",
          "endsOn",
          "count",
          "timezone",
          "materializedThrough",
          "startTimeLocal",
          "durationMinutes"
        ]
      },
      "LabelResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "householdId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "householdId",
          "name",
          "color",
          "createdBy",
          "createdAt"
        ]
      },
      "EventResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "householdId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "object"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "allDay": {
            "type": "boolean"
          },
          "location": {
            "type": "object"
          },
          "recurrenceRuleId": {
            "type": "object",
            "nullable": true
          },
          "occurrenceDate": {
            "type": "object",
            "nullable": true
          },
          "cancelledAt": {
            "type": "object",
            "nullable": true
          },
          "recurrence": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/RecurrenceResponseDto"
              }
            ]
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabelResponseDto"
            }
          }
        },
        "required": [
          "id",
          "householdId",
          "title",
          "startTime",
          "endTime",
          "allDay",
          "recurrenceRuleId",
          "occurrenceDate",
          "cancelledAt",
          "recurrence",
          "createdBy",
          "createdAt",
          "updatedAt",
          "labels"
        ]
      },
      "EventListResponseDto": {
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "materializedThrough": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "events",
          "total",
          "materializedThrough"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "expectedUpdatedAt": {
            "type": "string",
            "description": "Required version read before editing; stale writes return 409 EDIT_CONFLICT."
          },
          "expectedRuleUpdatedAt": {
            "type": "string",
            "description": "Required when editing a recurring occurrence."
          },
          "labelIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string",
            "description": "Event title",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "description": "Optional description"
          },
          "startTime": {
            "type": "string",
            "description": "Start time in ISO 8601 format"
          },
          "endTime": {
            "type": "string",
            "description": "End time in ISO 8601 format"
          },
          "allDay": {
            "type": "boolean",
            "description": "Whether this is an all-day event"
          },
          "location": {
            "type": "string",
            "description": "Optional location",
            "maxLength": 255
          },
          "recurrence": {
            "description": "重复规则；省略即不改变当前重复设置",
            "allOf": [
              {
                "$ref": "#/components/schemas/RecurrenceDto"
              }
            ]
          }
        },
        "required": [
          "expectedUpdatedAt"
        ]
      },
...
    }
  }
}